chore: resolve merge conflicts in feature/frontendGithubUsernames
